feat(app+views): Move logic from app to views components
