+++ projects/knora-ui/src/lib/viewer/index.ts
@@ -1,2 +1,3 @@
 export * from './viewer.module';
 export * from './values/index';
+export * from './views/resource-view/resource-view.component';
