+++ projects/knora-ui/src/lib/viewer/views/property-view/property-view.component.html
@@ -1,1 +1,13 @@
-<p>property-view works!</p>
+<div *ngFor="let prop of propArray">
+    <div *ngIf="prop.values">
+
+        <h4 class="label">
+            {{prop.propDef.label}}
+        </h4>
+
+        <div *ngFor="let val of prop.values">
+            <kui-display-edit #displayEdit [parentResource]="resource" [displayValue]="val"></kui-display-edit>
+        </div>
+
+    </div>
+</div>
